design: Confirm Modal 디자인 구현
